requirements.txt
pandas
numpy
requests
transformers
torch
scikit-learn
matplotlib
scipy
seaborn
pyyaml